Remove obsolete test module.
